+++ node/src/GlideClusterClient.ts
@@ -13,6 +13,7 @@ import {
     ReturnType,
 } from "./BaseClient";
 import {
+    DecoderOption,
     FlushMode,
     FunctionListOptions,
     FunctionListResponse,
@@ -691,8 +692,8 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async copy(
-        source: string,
-        destination: string,
+        source: GlideString,
+        destination: GlideString,
         replace?: boolean,
     ): Promise<boolean> {
         return this.createWritePromise(
@@ -1147,7 +1148,7 @@ export class GlideClusterClient extends BaseClient {
      * @see {@link https://valkey.io/commands/sort/|valkey.io} for details.
      *
      * @param key - The key of the list, set, or sorted set to be sorted.
-     * @param options - (Optional) {@link SortClusterOptions}.
+     * @param options - (Optional) {@link SortClusterOptions} and {@link DecoderOption}.
      * @returns An `Array` of sorted elements.
      *
      * @example
@@ -1158,10 +1159,12 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async sort(
-        key: string,
-        options?: SortClusterOptions,
-    ): Promise<string[]> {
-        return this.createWritePromise(createSort(key, options));
+        key: GlideString,
+        options?: SortClusterOptions & DecoderOption,
+    ): Promise<GlideString[]> {
+        return this.createWritePromise(createSort(key, options), {
+            decoder: options?.decoder,
+        });
     }
 
     /**
@@ -1175,7 +1178,7 @@ export class GlideClusterClient extends BaseClient {
      * @remarks Since Valkey version 7.0.0.
      *
      * @param key - The key of the list, set, or sorted set to be sorted.
-     * @param options - (Optional) {@link SortClusterOptions}.
+     * @param options - (Optional) {@link SortClusterOptions} and {@link DecoderOption}.
      * @returns An `Array` of sorted elements
      *
      * @example
@@ -1186,10 +1189,12 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async sortReadOnly(
-        key: string,
-        options?: SortClusterOptions,
-    ): Promise<string[]> {
-        return this.createWritePromise(createSortReadOnly(key, options));
+        key: GlideString,
+        options?: SortClusterOptions & DecoderOption,
+    ): Promise<GlideString[]> {
+        return this.createWritePromise(createSortReadOnly(key, options), {
+            decoder: options?.decoder,
+        });
     }
 
     /**
@@ -1218,8 +1223,8 @@ export class GlideClusterClient extends BaseClient {
      * ```
      */
     public async sortStore(
-        key: string,
-        destination: string,
+        key: GlideString,
+        destination: GlideString,
         options?: SortClusterOptions,
     ): Promise<number> {
         return this.createWritePromise(createSort(key, options, destination));
@@ -1234,6 +1239,7 @@ export class GlideClusterClient extends BaseClient {
      * @param route - (Optional) The command will be routed to a random node, unless `route` is provided, in which
      *     case the client will route the command to the nodes defined by `route`.
      * @returns `UNIX TIME` of the last DB save executed with success.
+     *
      * @example
      * ```typescript
      * const timestamp = await client.lastsave();
@@ -1252,7 +1258,9 @@ export class GlideClusterClient extends BaseClient {
      * @see {@link https://valkey.io/commands/randomkey/|valkey.io} for details.
      *
      * @param route - (Optional) The command will be routed to all primary nodes, unless `route` is provided,
-     *      in which case the client will route the command to the nodes defined by `route`.
+     *     in which case the client will route the command to the nodes defined by `route`.
+     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
+     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
      * @returns A random existing key name.
      *
      * @example
@@ -1261,9 +1269,13 @@ export class GlideClusterClient extends BaseClient {
      * console.log(result); // Output: "key12" - "key12" is a random existing key name.
      * ```
      */
-    public async randomKey(route?: Routes): Promise<string | null> {
+    public async randomKey(options?: {
+        route?: Routes;
+        decoder?: Decoder;
+    }): Promise<GlideString | null> {
         return this.createWritePromise(createRandomKey(), {
-            route: toProtobufRoute(route),
+            route: toProtobufRoute(options?.route),
+            decoder: options?.decoder,
         });
     }