+++ node/tests/SharedTests.ts
@@ -235,7 +235,11 @@ export function runBaseTests<Context>(config: {
                 expect(result).toEqual("OK");
                 result = await client.set(key3, value);
                 expect(result).toEqual("OK");
-                let deletedKeysNum = await client.del([key1, key2, key3]);
+                let deletedKeysNum = await client.del([
+                    key1,
+                    Buffer.from(key2),
+                    key3,
+                ]);
                 expect(deletedKeysNum).toEqual(3);
                 deletedKeysNum = await client.del([uuidv4()]);
                 expect(deletedKeysNum).toEqual(0);
@@ -3231,7 +3235,11 @@ export function runBaseTests<Context>(config: {
                 expect(await client.exists([key1])).toEqual(1);
                 expect(await client.set(key2, value)).toEqual("OK");
                 expect(
-                    await client.exists([key1, "nonExistingKey", key2]),
+                    await client.exists([
+                        key1,
+                        "nonExistingKey",
+                        Buffer.from(key2),
+                    ]),
                 ).toEqual(2);
                 expect(await client.exists([key1, key1])).toEqual(2);
             }, protocol);
@@ -3251,7 +3259,12 @@ export function runBaseTests<Context>(config: {
                 expect(await client.set(key2, value)).toEqual("OK");
                 expect(await client.set(key3, value)).toEqual("OK");
                 expect(
-                    await client.unlink([key1, key2, "nonExistingKey", key3]),
+                    await client.unlink([
+                        key1,
+                        key2,
+                        "nonExistingKey",
+                        Buffer.from(key3),
+                    ]),
                 ).toEqual(3);
             }, protocol);
         },
@@ -3272,26 +3285,32 @@ export function runBaseTests<Context>(config: {
                     cluster.checkIfServerVersionLessThan("7.0.0");
 
                 if (versionLessThan) {
-                    expect(await client.pexpire(key, 10000)).toEqual(true);
+                    expect(
+                        await client.pexpire(Buffer.from(key), 10000),
+                    ).toEqual(true);
                 } else {
                     expect(
                         await client.pexpire(
-                            key,
+                            Buffer.from(key),
                             10000,
                             ExpireOptions.HasNoExpiry,
                         ),
                     ).toEqual(true);
                 }
 
-                expect(await client.ttl(key)).toBeLessThanOrEqual(10);
+                expect(await client.ttl(Buffer.from(key))).toBeLessThanOrEqual(
+                    10,
+                );
 
                 /// TTL will be updated to the new value = 15
                 if (versionLessThan) {
-                    expect(await client.expire(key, 15)).toEqual(true);
+                    expect(await client.expire(Buffer.from(key), 15)).toEqual(
+                        true,
+                    );
                 } else {
                     expect(
                         await client.expire(
-                            key,
+                            Buffer.from(key),
                             15,
                             ExpireOptions.HasExistingExpiry,
                         ),
@@ -3302,6 +3321,12 @@ export function runBaseTests<Context>(config: {
                     expect(await client.pexpiretime(key)).toBeGreaterThan(
                         Date.now(),
                     );
+                    expect(
+                        await client.expiretime(Buffer.from(key)),
+                    ).toBeGreaterThan(Math.floor(Date.now() / 1000));
+                    expect(
+                        await client.pexpiretime(Buffer.from(key)),
+                    ).toBeGreaterThan(Date.now());
                 }
 
                 expect(await client.ttl(key)).toBeLessThanOrEqual(15);
@@ -3329,14 +3354,14 @@ export function runBaseTests<Context>(config: {
                 if (versionLessThan) {
                     expect(
                         await client.expireAt(
-                            key,
+                            Buffer.from(key),
                             Math.floor(Date.now() / 1000) + 50,
                         ),
                     ).toEqual(true);
                 } else {
                     expect(
                         await client.expireAt(
-                            key,
+                            Buffer.from(key),
                             Math.floor(Date.now() / 1000) + 50,
                             ExpireOptions.NewExpiryGreaterThanCurrent,
                         ),
@@ -3352,6 +3377,13 @@ export function runBaseTests<Context>(config: {
                     expect(
                         await client.pexpireAt(
                             key,
+                            Date.now() + 50000,
+                            ExpireOptions.HasExistingExpiry,
+                        ),
+                    ).toEqual(false);
+                    expect(
+                        await client.pexpireAt(
+                            Buffer.from(key),
                             Date.now() + 50000,
                             ExpireOptions.HasExistingExpiry,
                         ),
@@ -4775,7 +4807,7 @@ export function runBaseTests<Context>(config: {
                 expect(await client.del([key])).toEqual(1);
 
                 expect(await client.lpush(key, ["value"])).toEqual(1);
-                expect(await client.type(key)).toEqual("list");
+                expect(await client.type(Buffer.from(key))).toEqual("list");
                 expect(await client.del([key])).toEqual(1);
 
                 expect(await client.sadd(key, ["value"])).toEqual(1);
@@ -4787,7 +4819,7 @@ export function runBaseTests<Context>(config: {
                 expect(await client.del([key])).toEqual(1);
 
                 expect(await client.hset(key, { field: "value" })).toEqual(1);
-                expect(await client.type(key)).toEqual("hash");
+                expect(await client.type(Buffer.from(key))).toEqual("hash");
                 expect(await client.del([key])).toEqual(1);
 
                 await client.xadd(key, [["field", "value"]]);
@@ -5069,7 +5101,7 @@ export function runBaseTests<Context>(config: {
                 expect(await client.pttl(key)).toEqual(-1);
 
                 expect(await client.expire(key, 10)).toEqual(true);
-                let result = await client.pttl(key);
+                let result = await client.pttl(Buffer.from(key));
                 expect(result).toBeGreaterThan(0);
                 expect(result).toBeLessThanOrEqual(10000);
 
@@ -5274,7 +5306,7 @@ export function runBaseTests<Context>(config: {
                 expect(await client.persist(key)).toEqual(false);
 
                 expect(await client.expire(key, 10)).toEqual(true);
-                expect(await client.persist(key)).toEqual(true);
+                expect(await client.persist(Buffer.from(key))).toEqual(true);
             }, protocol);
         },
         config.timeout,
@@ -6193,11 +6225,15 @@ export function runBaseTests<Context>(config: {
                 const key = uuidv4() + "{123}";
                 const newKey = uuidv4() + "{123}";
                 await client.set(key, "value");
-                await client.rename(key, newKey);
-                const result = await client.get(newKey);
-                expect(result).toEqual("value");
+                expect(await client.rename(key, newKey)).toEqual("OK");
+                expect(await client.get(newKey)).toEqual("value");
                 // If key doesn't exist it should throw, it also test that key has successfully been renamed
                 await expect(client.rename(key, newKey)).rejects.toThrow();
+                // rename back
+                expect(
+                    await client.rename(Buffer.from(newKey), Buffer.from(key)),
+                ).toEqual("OK");
+                expect(await client.get(key)).toEqual("value");
             }, protocol);
         },
         config.timeout,
@@ -6222,11 +6258,15 @@ export function runBaseTests<Context>(config: {
                 await client.set(key1, "key1");
                 await client.set(key3, "key3");
                 // Test that renamenx can rename key1 to key2 (non-existing value)
-                expect(await client.renamenx(key1, key2)).toEqual(true);
+                expect(await client.renamenx(Buffer.from(key1), key2)).toEqual(
+                    true,
+                );
                 // sanity check
                 expect(await client.get(key2)).toEqual("key1");
                 // Test that renamenx doesn't rename key2 to key3 (with an existing value)
-                expect(await client.renamenx(key2, key3)).toEqual(false);
+                expect(await client.renamenx(key2, Buffer.from(key3))).toEqual(
+                    false,
+                );
                 // sanity check
                 expect(await client.get(key3)).toEqual("key3");
             }, protocol);
@@ -6808,7 +6848,9 @@ export function runBaseTests<Context>(config: {
                 expect(await client.objectEncoding(string_key)).toEqual("raw");
 
                 expect(await client.set(string_key, "2")).toEqual("OK");
-                expect(await client.objectEncoding(string_key)).toEqual("int");
+                expect(
+                    await client.objectEncoding(Buffer.from(string_key)),
+                ).toEqual("int");
 
                 expect(await client.set(string_key, "value")).toEqual("OK");
                 expect(await client.objectEncoding(string_key)).toEqual(
@@ -6899,7 +6941,9 @@ export function runBaseTests<Context>(config: {
 
                 if (versionLessThan7) {
                     expect(
-                        await client.objectEncoding(zset_listpack_key),
+                        await client.objectEncoding(
+                            Buffer.from(zset_listpack_key),
+                        ),
                     ).toEqual("ziplist");
                 } else {
                     expect(
@@ -6938,9 +6982,9 @@ export function runBaseTests<Context>(config: {
                         null,
                     );
                     expect(await client.set(key, "foobar")).toEqual("OK");
-                    expect(await client.objectFreq(key)).toBeGreaterThanOrEqual(
-                        0,
-                    );
+                    expect(
+                        await client.objectFreq(Buffer.from(key)),
+                    ).toBeGreaterThanOrEqual(0);
                 } finally {
                     expect(
                         await client.configSet({
@@ -6977,7 +7021,9 @@ export function runBaseTests<Context>(config: {
 
                     await wait(2000);
 
-                    expect(await client.objectIdletime(key)).toBeGreaterThan(0);
+                    expect(
+                        await client.objectIdletime(Buffer.from(key)),
+                    ).toBeGreaterThan(0);
                 } finally {
                     expect(
                         await client.configSet({
@@ -7005,9 +7051,9 @@ export function runBaseTests<Context>(config: {
 
                 expect(await client.objectRefcount(nonExistingKey)).toBeNull();
                 expect(await client.set(key, "foo")).toEqual("OK");
-                expect(await client.objectRefcount(key)).toBeGreaterThanOrEqual(
-                    1,
-                );
+                expect(
+                    await client.objectRefcount(Buffer.from(key)),
+                ).toBeGreaterThanOrEqual(1);
             }, protocol);
         },
         config.timeout,
@@ -8331,7 +8377,9 @@ export function runBaseTests<Context>(config: {
                 expect(
                     await client.mset({ [key1]: "value1", [key2]: "value2" }),
                 ).toEqual("OK");
-                expect(await client.touch([key1, key2])).toEqual(2);
+                expect(await client.touch([key1, Buffer.from(key2)])).toEqual(
+                    2,
+                );
                 expect(
                     await client.touch([key2, nonExistingKey, key1]),
                 ).toEqual(2);